refactor/create hook useScreenWidth and debug/no redundand redirect
